chore: Update Next.js configuration files and enable server components
Update the Next.js configuration files to include the necessary changes for enabling server components. This includes adding the "missingSuspenseWithCSRBailout" option with a value of false. Additionally, update the "serverComponentsExternalPackages" array to include "mongoose". This ensures that the server components work properly and that the "mongoose" package is available externally.

Refactor the CSS styles for the copy button and glassmorphism components to improve consistency and readability. Update the class names and apply consistent formatting for background color, shadow, and border properties.

Update project dependencies to their latest versions and add Next.js configuration files for improved project setup and performance.